win64: use int since size_t is 64-bits and breaks wxToString

+++ src/effects/Equalization.h
@@ -193,7 +193,7 @@ private:
    float *mFFTBuffer;
    float *mFilterFuncR;
    float *mFilterFuncI;
-   size_t mM;
+   int mM;
    wxString mCurveName;
    bool mLin;
    float mdBMax;
